third_party/blink/web_tests/svg/custom/use-property-synchronization-crash-expected.txt

CONSOLE ERROR: line 5: Error: <polyline> attribute points: Unexpected end of attribute. Expected number, "0".
CONSOLE ERROR: line 11: Error: <polyline> attribute points: Unexpected end of attribute. Expected number, "0".
